Update envoy to 11/03 (#3085)
Head of istio/istio has a bug causing crashes, head of istio/proxy has a
bug causing pods to not startup. Both blocking tons of PRs in
istio/istio.

This commit has a fix for SDS reverted, but it impacts only real large
scale clusters and rarely CI, so lets pull it in now to unblock